% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.R
\docType{data}
\name{expr}
\alias{expr}
\title{Gene expression data for 2 rat tissues}
\format{A data list that containing 2 data matrices
\describe{
  \item{brain}{expression levels of 100 gene for brain tissue}
  \item{heart}{expression levels of 100 gene for heart tissue}
}}
\source{
More detailed information can be found in [PhenoGen website](http://phenogen.ucdenver.edu).
}
\usage{
expr
}
\description{
A dataset containing gene expression levels for rat brain and heart tissues. The expression intensity were measured using Affymetrix Rat Exon 1.0 ST Array. 100 differential genes are selected to construct example co-expression network for 2 tissues.
}
\keyword{datasets}
