% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_downstream.R
\name{get_sp_net_2t}
\alias{get_sp_net_2t}
\title{Condition-adaptive fused graphical lasso}
\usage{
get_sp_net_2t(rmat)
}
\arguments{
\item{Y}{A list expression data which are n*p matrices. all matrices should have a same n and p.}

\item{lambda1}{The tuning parameter for the graphical lasso penalty.}

\item{lambda2}{The tuning parameter for the between condition group lasso penalty.}
}
\value{
\code{CFGL} produces a list that contains estimated inverse matrices and other necessary components.
}
\description{
The function jointly construct gene co-expression network for multiple class using Condition-adaptive Fused Graphical Lasso. Pairwise screening matrics are required to adjust between-condition lasso penalty.
}
